[ALPS04140405] mdla: fix build warnings
fix build warnings that converted to errors:

mdla_main.c:3946:21: error: cast from pointer to integer of
  different size
mdla_main.c:1561:6: error: "ret" may be used uninitialized
  in this function
mdla_proc.c:71:2: error: ignoring return value of
  "copy_to_user", declared with attribute warn_unused_result

... and more.

MTK-Commit-Id: 73986586067269c9042ce220576b24238f32355c

CR-Id: ALPS04140405
Feature: MDLA Platform

+++ drivers/misc/mediatek/mdla/1.0/mdla_trace.c
@@ -115,7 +115,7 @@ void mdla_dump_prof(struct seq_file *s)
 	int i;
 
 #define _SHOW_VAL(t) \
-	mdla_print_seq(s, "%s=%d\n", #t, cfg_##t)
+	mdla_print_seq(s, "%s=%lu\n", #t, (unsigned long)cfg_##t)
 
 	_SHOW_VAL(period);
 	_SHOW_VAL(cmd_trace);
